+++ examples/plugin/src/index.ts
@@ -1,7 +1,7 @@
 import { BundleTaskConfig, TaskName } from '@ice/pkg';
-import type { PkgPlugin } from '@ice/pkg';
+import type { Plugin } from '@ice/pkg';
 
-const plugin: PkgPlugin = (api) => {
+const plugin: Plugin = (api) => {
   const { onGetConfig } = api;
 
   const bundleTaskCallback: Parameters<typeof onGetConfig>[0] = async (config: BundleTaskConfig) => {
@@ -23,14 +23,13 @@ const plugin: PkgPlugin = (api) => {
     // config.sourcemap = true;
     config.alias = { ...config.alias };
     config.externals = { react: 'React', 'react-dom': 'ReactDOM' };
-    // config.bundle.filename = ({ format, development, isES2017 }) => {
-    //   return `index${(format === 'umd' || format === 'cjs') ? `.${format}` : ''}${isES2017 ? '-2017' : ''}${development ? 'development' : ''}.js`;
-    // };
-    config.stylesOptions = (options) => {
+
+    config.modifyStylesOptions ??= [];
+    config.modifyStylesOptions.push((options) => {
       return options;
-    };
-    return config;
+    });
   };
+
   onGetConfig(TaskName.BUNDLE_ES2017, bundleTaskCallback);
   onGetConfig(TaskName.BUNDLE_ES5, bundleTaskCallback);
 };