+++ packages/pkg/src/types.ts
@@ -1,22 +1,20 @@
-import type { RollupOptions, Plugin as RollupPlugin, SourceMapInput, ModuleJSON, RollupOutput } from 'rollup';
-import type { Context, PluginAPI, Plugin, TaskConfig as BuildTask } from 'build-scripts';
+import type { RollupOptions, SourceMapInput, ModuleJSON, RollupOutput } from 'rollup';
+import type { Context as _Context, PluginAPI as _PluginAPI, Plugin as _Plugin, TaskConfig as _BuildTask } from 'build-scripts';
 import type { Config } from '@swc/core';
 import type stylesPlugin from 'rollup-plugin-styles';
+import type { FSWatcher } from 'chokidar';
 
-type StylesRollupPluginOptions = Parameters<typeof stylesPlugin>[0];
+export type StylesRollupPluginOptions = Parameters<typeof stylesPlugin>[0];
 
-export type PlainObject = Record<string, string | boolean | number | object>;
-export type ReverseMap<T> = T[keyof T];
+export type PlainObject = Record<string, string | boolean | number | object | null>;
 
-export type RollupPluginFn<T = {}> = (args?: T) => RollupPlugin;
-
-interface TransformOptions {
+export interface TransformUserConfig {
   /**
    * Which type of contents would be generated
    * "cjs" - Commonjs with ES5 syntax (targeting Node version under 12);
    * "esm" - ES Module with ES5 syntax (legacy outputs);
    * "es2017" - ES Module with ES2017 (targeting modern browsers and Node version upon 12)
-   * @default ['esm','es2017']
+   * @default ['esm', 'es2017']
    */
   formats?: Array<'cjs' | 'esm' | 'es2017'>;
   /**
@@ -28,25 +26,32 @@ interface TransformOptions {
   excludes?: string | string[];
 }
 
-interface BundleOptions {
+export interface BundleUserConfig {
   /**
    * Export name
    * @default package.name
    */
   name?: string;
   /**
-   * As the name of the generated file.
-   * @default index
+   * Output directory
+   * @default dist
    */
-  filename?: string;
+  outputDir?: string;
   /**
+   * @deprecated Please use `bundle.modes` config.
    * Generate uncompressed bundle for development debug.
    */
   development?: boolean;
   /**
+   * Node env modes. For example: 'production', 'development'
+   * @default ['production']
+   */
+  modes?: NodeEnvMode[];
+  /**
    * Which type of contents would be generated
    * "umd"
    * "esm"
+   * "cjs"
    * "es2017"
    * @default ['esm','es2017']
    */
@@ -65,23 +70,56 @@ interface BundleOptions {
   minify?: boolean;
 }
 
-export type TaskLoaderConfig = BundleTaskLoaderConfig | TransformTaskLoaderConfig;
-
-export interface BundleTaskLoaderConfig extends BundleTaskConfig {
-  type: 'bundle';
-  name: TaskName;
-}
-
-export interface TransformTaskLoaderConfig extends TransformTaskConfig {
-  type: 'transform';
-  name: TaskName;
-}
-
-interface CommonTaskConfig {
+export interface UserConfig {
   /**
-   * Build mode.
+  * Entry for a task
+  * @default  `./src/index`
+  */
+  entry?: RollupOptions['input'];
+  /**
+   * Alias to file system paths
+   * @default {}
    */
-  mode?: NodeEnvMode;
+  alias?: Record<string, string>;
+  /**
+   * Define global constant replacements
+   */
+  define?: PlainObject;
+  /**
+  * - true to generate a sourcemap for the code and include it in the result object.
+  * - "inline" to generate a sourcemap and append it as a data URL to the end of the code,
+  * but not include it in the result object.
+   */
+  sourceMaps?: boolean | 'inline';
+  /**
+  * Whether or not to generate declaration files for Ecmascript
+  * @default false
+  */
+  generateTypesForJs?: boolean;
+
+  /**
+   * Plugins of build scripts
+   * @default []
+   */
+  plugins?: Array<string | [string, any?]>;
+
+  /**
+   * "transform mode" means transform files one by one
+   */
+  transform?: TransformUserConfig;
+
+  /**
+   * "bundle mode" means bundle everything up by using Rollup
+   */
+  bundle?: BundleUserConfig;
+}
+
+interface _TaskConfig {
+  /**
+  * Entry for a task
+  * @default  `./src/index`
+  */
+  entry?: RollupOptions['input'];
   /**
   * Output directory
   */
@@ -89,7 +127,7 @@ interface CommonTaskConfig {
   /**
    * Define global constant replacements
    */
-  define?: Record<string, string>;
+  define?: PlainObject;
   /**
     * - true to generate a sourcemap for the code and include it in the result object.
     * - "inline" to generate a sourcemap and append it as a data URL to the end of the code,
@@ -101,14 +139,10 @@ interface CommonTaskConfig {
    */
   alias?: Record<string, string>;
   /**
-  * Extra rollup options
-  * @see https://rollupjs.org/guide/en/
-  */
-  rollupOptions?: RollupOptions;
-  /**
-  * Extra rollup plugins
-  */
-  rollupPlugins?: RollupPlugin[];
+   * Modify default rollup options
+   * @see https://rollupjs.org/guide/en/#rolluprollup
+   */
+  modifyRollupOptions?: Array<(rollupOptions: RollupOptions) => RollupOptions>;
   /**
   * Extra swc compile options
   * @see https://swc.rs/docs/configuration/compilationv
@@ -120,13 +154,10 @@ interface CommonTaskConfig {
   babelPlugins?: babel.PluginItem[];
 }
 
-export interface BundleTaskConfig extends CommonTaskConfig, BundleOptions {
+export interface BundleTaskConfig extends
+  _TaskConfig,
+  Omit<BundleUserConfig, 'development'> {
   type: 'bundle';
-  /**
-  * Entry for a specific task
-  * @default  `./src/index[j|t]sx` for bundling task
-  */
-  entry?: RollupOptions['input'];
   /**
   * Files extensions
   * @see https://www.npmjs.com/package/@rollup/plugin-node-resolve
@@ -135,28 +166,37 @@ export interface BundleTaskConfig extends CommonTaskConfig, BundleOptions {
   /**
    * Config styles options. See https://www.npmjs.com/package/rollup-plugin-styles
    */
-  stylesOptions?: (options: StylesRollupPluginOptions) => StylesRollupPluginOptions;
+  modifyStylesOptions?: Array<(options: StylesRollupPluginOptions) => StylesRollupPluginOptions>;
 }
 
-export interface TransformTaskConfig extends CommonTaskConfig, TransformOptions {
+export interface TransformTaskConfig extends _TaskConfig, TransformUserConfig {
   type: 'transform';
   /**
-  * Entry for a specific task
-  * @default  `./src` for Transform task
-  */
-  entry?: string;
+   * Node env modes. For example: 'production', 'development'
+   * @default `['development']` on start, `['production']` on build.
+   */
+  modes?: NodeEnvMode[];
 }
 
 export type TaskConfig = BundleTaskConfig | TransformTaskConfig;
 
-export type PkgTaskConfig = BuildTask<TaskConfig, TaskName>;
+export type BuildTask = _BuildTask<TaskConfig, TaskName>;
 
-export type PkgContext = Context<TaskConfig, {}, UserConfig>;
+export type Context = _Context<TaskConfig, {}, UserConfig>;
 
-export type PkgPluginAPI = PluginAPI<TaskConfig>;
+export type PluginAPI = _PluginAPI<TaskConfig>;
+/**
+ * @deprecated Please use PluginAPI instead.
+ */
+export type PkgPluginAPI = PluginAPI;
 
-export type PkgPlugin = Plugin<TaskConfig>;
+export type Plugin = _Plugin<TaskConfig>;
+/**
+ * @deprecated Please use Plugin instead.
+ */
+export type PkgPlugin = Plugin;
 
+// TODO: The enum name should be renamed to Task.
 export enum TaskName {
   'TRANSFORM_CJS' = 'transform-cjs',
   'TRANSFORM_ESM' = 'transform-esm',
@@ -164,6 +204,9 @@ export enum TaskName {
   'BUNDLE_ES5' = 'bundle-es5',
   'BUNDLE_ES2017' = 'bundle-es2017',
 }
+type TaskKey = keyof typeof TaskName;
+// TODO: The type name should be renamed to TaskName.
+export type TaskValue = typeof TaskName[TaskKey];
 
 export interface OutputFile {
   // globby parsed path, which is relative
@@ -190,48 +233,23 @@ export interface OutputResult {
   outputs?: Array<RollupOutput['output']>;
 }
 
-export interface UserConfig {
-  /**
-   * Entry for build.
-   */
-  entry?: string | string[] | Record<string, string>;
-  /**
-   * Alias to file system paths
-   * @default {}
-   */
-  alias?: Record<string, string>;
-  /**
-   * Define global constant replacements
-   */
-  define?: Record<string, string>;
-  /**
-  * - true to generate a sourcemap for the code and include it in the result object.
-  * - "inline" to generate a sourcemap and append it as a data URL to the end of the code,
-  * but not include it in the result object.
-   */
-  sourceMaps?: boolean | 'inline';
-  /**
-  * Whether or not to generate declation files for Ecmascript
-  * @default false
-  */
-  generateTypesForJs?: boolean;
-
-  /**
-   * Plugins of build scripts
-   * @default []
-   */
-  plugins?: Array<string | [string, any?]>;
-
-  /**
-   * "transform mode" means transform files one by one
-   */
-  transform?: TransformOptions;
-
-  /**
-   * "bundle mode" means bundle everything up by using Rollup
-   */
-  bundle?: BundleOptions;
-}
-
 // Set for `process.env.NODE_ENV`
 export type NodeEnvMode = 'development' | 'production' | string;
+
+export type WatchEvent = 'create' | 'update' | 'delete';
+export type HandleChange<R = OutputResult> = (id: string, event: WatchEvent) => Promise<R>;
+
+export interface TaskResult {
+  handleChange?: HandleChange<OutputResult[]>;
+  outputResults: OutputResult[];
+}
+export interface TaskRunnerContext {
+  mode: NodeEnvMode;
+  buildTask: BuildTask;
+}
+
+export type RunTasks = (
+  taskOptionsList: Array<[RollupOptions, TaskRunnerContext]>,
+  context: Context,
+  watcher?: FSWatcher,
+) => Promise<TaskResult>;